requirements.txt
# 核心大模型统一接入层，支持100+大模型，自带重试、异常处理、异步支持
litellm>=1.40.0
# 可视化面板
streamlit>=1.30.0
# 配置文件管理
pyyaml>=6.0
python-dotenv>=1.0.0
# 异步与并发
aiohttp>=3.9.0
# 时间与日期
python-dateutil>=2.8.0
# 日志与调试
loguru>=0.7.0
# 数据处理
pandas>=2.0.0
# 网络请求
requests>=2.31.0
# 跨进程文件锁
filelock>=3.13.0
# LLM API客户端
openai>=1.0.0
anthropic>=0.25.0
# Notion客户端
notion-client>=2.0.0